python/requirements.txt
numpy
tqdm
cupy-cuda12x
